% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_multiomic_relation.R
\name{analyze_multiomic_relation}
\alias{analyze_multiomic_relation}
\title{analyze_multiomic_relation}
\usage{
analyze_multiomic_relation(
  omics_list,
  multiomic_relation,
  test = "t-test",
  inter.omics.cor = FALSE,
  pval_threshold = 0.05
)
}
\arguments{
\item{omics_list}{A named list of omics matrices (each matrix: rows = features, columns = samples).}

\item{multiomic_relation}{multiomic_relation object with sample_ids: list mapping omic names to vector of sample IDs in biclusters and features: list mapping omic names to vector of features in biclusters.}

\item{test}{A test type: "t-test" (default) or "wilcox".}

\item{inter.omics.cor}{Logical. TRUE if feature-by-feature correlation is desired.}

\item{pval_threshold}{Default is 0.05. Only needed when inter.omics.cor is TRUE}
}
\value{
A data.frame with feature name, omic name, estimate (mean difference), p-value.
}
\description{
Performs a two-group test on each feature between bicluster and out-of-bicluster samples.
}
\examples{
analyze_multiomic_relation()
}
