% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_eigenfeatures.R
\name{correlate_eigenfeatures}
\alias{correlate_eigenfeatures}
\title{correlate_eigenfeatures}
\usage{
correlate_eigenfeatures(
  omic_list,
  multiomic_relation,
  output_prefix = "eigen_cor_plot"
)
}
\arguments{
\item{omic_list}{A named list of omic data matrices (genes, CpGs, etc.). Each matrix has features (rows) and samples (columns).}

\item{multiomic_relation}{Multiomic bicluster relations. Output from 'find_relations'}

\item{output_prefix}{Prefix for output name. For example, "eigen_cor_plot"}
}
\value{
List object with correlation matrix, p-value matrix and pairwise scatter plots.
}
\description{
Computes a correlation matrix between the summary expression levels (PC1s) of different omic biclusters.
}
\examples{
correlate_eigenfeatures()
}
