% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosbi_wrapper.R
\name{find_ensemble_bc}
\alias{find_ensemble_bc}
\title{find_ensemble_bc}
\usage{
find_ensemble_bc(data_matrix)
}
\arguments{
\item{data_matrix}{Omic data matrix with feature as rows and samples as columns}
}
\value{
List of ensemble biclusters.
}
\description{
Wrapper function for running mosbi
}
\examples{
find_ensemble_bc()
}
