% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_relations.R
\name{find_relations}
\alias{find_relations}
\title{find_relations}
\usage{
find_relations(biclusters_list, module, pheno, bc_dist_obj, bc_hclust_obj)
}
\arguments{
\item{biclusters_list}{A named list of biclusters for all omics. Output from 'find_ensemble_bc'. Must be named as c("gwe","meth","prot").}

\item{module}{A vector containing member bicluster names}

\item{pheno}{Data frame with sample IDs as one of the columns with name 'sampleID'}

\item{bc_dist_obj}{Output from 'generate_jaccard_dist' for bicluster matrix and jaccard dist for sample ids and bicluster names respectively}

\item{bc_hclust_obj}{Output from 'hclust_bcs' function for module names/colours}
}
\value{
The best multi-omic relations based on sample overlap count.
}
\description{
Finds all possible combinations across the input omic biclusters, calculates overlaps for each multi-omic set and return the best overlapping combination.
}
\examples{
find_relations()
}
