% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_jaccard_dist.R
\name{generate_jaccard_dist}
\alias{generate_jaccard_dist}
\title{generate_jaccard_dist}
\usage{
generate_jaccard_dist(mosbi_outputs, omic_names, omic_data)
}
\arguments{
\item{mosbi_outputs}{List of mosbi biclustering outputs for individual omics}

\item{omic_names}{Names of the omics in the same order (eg, c(meth,expr,protein))}

\item{omic_data}{List of omic data in same order.}
}
\value{
Jaccard distance between the biclusters across all omics.
}
\description{
Calculates Jaccard distance between the biclusters across all omics.
}
\examples{
generate_jaccard_dist()
}
