% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_bcs.R
\name{hclust_bcs}
\alias{hclust_bcs}
\title{hclust_bcs}
\usage{
hclust_bcs(jaccard_dist, plots = TRUE, cutHeight = 0.8, minModuleSize = 3)
}
\arguments{
\item{jaccard_dist}{This is analogous to dissTOM in standard WGCNA pipeline where columns are biclusterss}

\item{plots}{Logical (default:TRUE)}

\item{cutHeight}{Cut the tree at the height of distance (default:0.8)}

\item{minModuleSize}{Minimum number of biclusters in a module (default: 3)}
}
\value{
A list object containing a list of all modules, list of only multiomic modules and module names (colour) assigned to all biclusters
}
\description{
Performs hierarchical clustering of multi-omic biclusters using Jaccard index based distance
}
\examples{
hclust_bcs()
}
