% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_overlap_pval.R
\name{perm.overlap.sig}
\alias{perm.overlap.sig}
\title{perm.overlap.sig}
\usage{
perm.overlap.sig(omic_relations, sample_list)
}
\arguments{
\item{omic_relations}{Output from 'find_relations'}

\item{sample_list}{Vector of sample names in the omics data (all omics must have exact same sample list).}
}
\value{
P-value from prop.test
}
\description{
Calculates permutation based p-value for sample overlap among the biclusters from the different omics.
}
\examples{
risk_score_relation()
}
