% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_boxplot.R
\name{plot_feature_boxplot}
\alias{plot_feature_boxplot}
\title{plot_feature_boxplot}
\usage{
plot_feature_boxplot(
  omics_list,
  result_list,
  multiomic_relation,
  omic_type,
  rank = 1
)
}
\arguments{
\item{omics_list}{Named list of omic data matrices (features × samples).}

\item{result_list}{Output from analyze_multiomic_relation() (a list of data frames).}

\item{multiomic_relation}{multiomic_relation object}

\item{omic_type}{Name of the omic layer to plot (e.g., "expression", "methylation").}

\item{rank}{Which ranked feature to plot (e.g., 1 (default) for top feature by p-value).}
}
\value{
Boxplot
}
\description{
Plot boxplot to visualize distributional differences in feature expression/methylation/abundance between individuals in bicluster and those outside the bicluster.
}
\examples{
plot_feature_boxplot()
}
